sb_stemmer_stem takes a byte count not character count
This change *finally* means we actually handle Unicode correctly, and we can
rely on the UTF-8 interface completely, which in turn means we don't need to
depend on ICU.  This also turns out to be a lot faster!

The previous work-around didn't actually fix Unicode handling at all; it merely
masked the problem by using character sets where the relevant alphabets were
all in the single-byte range.