test thread-safety (currently failing!)
By stemming the test data with massive parallelism we should be able to
trigger any lacking thread-safety by having this test script crash.

This is currently failing because, as it turns out, neither
unsafeDupablePerformIO and withMVar nor lazy unsafeIOtoST are good
enough here.  The fix will be easy but impact performance.